feat(runtime): #82 extract recording infrastructure to recording.py
